@@ -1,37 +1,23 @@
+import { useState } from 'react';
+
 import { User } from './User';
 import { Login } from './Login';
 
-export class AppComponent extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      user: null
-    };
-    this.setUser = this.setUser.bind(this);
-  }
+export const AppComponent = () => {
+  const [user, setUser] = useState(null)
 
-  setUser(user) {
-    this.setState({
-      user
-    });
-  }
+  return (
+    <React.Fragment>
+      <h1>Simple react app</h1>
 
-  render() {
-    const { user } = this.state;
-
-    return (
-      <React.Fragment>
-        <h1>Simple react app</h1>
-
-        {
-          user ?
-            <User
-              data={user}
-              login={this.setUser}
-            /> :
-            <Login login={this.setUser}/>
-        }
-      </React.Fragment>
-    );
-  }
+      {
+        user ?
+          <User
+            data={user}
+            login={setUser}
+          /> :
+          <Login login={setUser} />
+      }
+    </React.Fragment>
+  )
 }
